Updates to leverage OOP design principles
- Create a ShapeBase to help reduce common code.
- Remove unused classes.
