+++ Shapes/Shapes/Rectangle.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// A Rectangle is a shape that has 4 strait sides and 4 right angles.
     /// </summary>
-    public class Rectangle
+    public class Rectangle : ShapeBase
     {
         /// <summary>
         /// Build a Rectangle object with a given length and a given height.
@@ -11,16 +11,10 @@
         /// <param name="length"><see cref="Length"/></param>
         /// <param name="height"></param>
         /// <summary>length represents one side of the rectangle where height represents a side that is perpendicular to the first side</summary>
-        public Rectangle(int length, int height) // Special type of method called a Constructor (same name as class).
+        public Rectangle(int length, int height, Color color = 0, Texture texture = 0) : base(length, color, texture)  // Special type of method called a Constructor (same name as class).
         {
-            Length = length;
             Height = height;
         }
-
-        /// <summary>
-        /// Size of a side of the Rectangle.
-        /// </summary>
-        public int Length { get; set; } // Example of a property.
         
         /// <summary>
         /// Side perpendicular to the Length of the Rectangle.
@@ -31,7 +25,7 @@
         /// Calculate the area of the Rectangle (length * height).
         /// </summary>
         /// <returns></returns>
-        public int Area() // Example of a method.
+        public override int Area() // Example of a method.
         {
             return Length * Height;
         }
